+++ pathBuilder.js
@@ -14,7 +14,7 @@ const buildPaths = config => {
         jsSrcDir: `${config.assetSrcDir}/${config.js.jsDir}`,
         jsDistDir: `${config.assetDistDir}/${config.js.jsDir}`,
 
-        jsErrorLoggerSubDir: `${config.webRootDir}/${config.logger.dir}/${config.logger.subDir}`,
+        jsErrorLoggerDir: `${config.webRootDir}/${config.logger.dir}`,
 
         imgSrcDir: `${config.assetSrcDir}/${config.img.imgDir}`,
         imgDistDir: `${config.assetDistDir}/${config.img.imgDir}`,